feat(lsp): Add LspAttached and LspDetached autocommands
The current approach of using `on_attach` callbacks for configuring
buffers for LSP is suboptimal:

1. It does not use the standard Nvim interface for driving and hooking
   into events (i.e. autocommands)
2. There is no way for "third parties" (e.g. plugins) to hook into the
   event. This means that *all* buffer configuration must go into the
   user-supplied on_attach callback. This also makes it impossible for
   these configurations to be modular, since it all must happen in the
   same place.
3. There is currently no way to do something when a client detaches from
   a buffer (there is no `on_detach` callback).

The solution is to use the traditional method of event handling in Nvim:
autocommands. When a LSP client is attached to a buffer, fire a
`LspAttached` event with the client's name as the pattern. Likewise,
when a client detaches from a buffer fire a `LspDetached` event.

This enables plugins to easily add LSP-specific configuration to buffers
as well as enabling users to make their own configurations more modular
(e.g. by creating multiple LspAttached autocommands that each do
something unique).

The new `vim.lsp.get_client_by_name()` function allows users to retrieve
the client object from the autocommand callback by using `<amatch>` or
the `match` field of the `args` table in Lua callbacks. In the future we
can explore passing the client object directly to Lua callbacks.